renamed slotsbootstrap to slots
