+++ reahl-doc/reahl/doc/examples/tutorial/slots/.reahlproject
@@ -17,6 +17,6 @@
     <egg name="reahl-webdev"/>
   </deps>
 
-  <alias name="unit" command="setup -- -q pytest --addopts '-s --pyargs reahl.doc.examples.tutorial.slotsbootstrap.slotsbootstrap_dev'"/>
+  <alias name="unit" command="setup -- -q pytest --addopts '-s --pyargs reahl.doc.examples.tutorial.slots.slots_dev'"/>
 
 </project>
