+++ reahl-doc/reahl/doc/examples/tutorial/slots/etc/web.config.py
@@ -1,5 +1,5 @@
 
 from __future__ import print_function, unicode_literals, absolute_import, division
-from reahl.doc.examples.tutorial.slotsbootstrap.slotsbootstrap import SlotsUI
+from reahl.doc.examples.tutorial.slots.slots import SlotsUI
 
 web.site_root = SlotsUI
